+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+## 1.0.9 (2022-05-05)
+
 - Add `AnyFixture.before_fixtures_reset` and `AnyFixture.after_fixtures_reset` callbacks. ([@ruslanshakirov][])
 
 - Fixes ActiveRecord 6.1 issue with AnyFixture and Postgres config ([@markedmondson][])